+++ tests/testthat/test-wc.R
@@ -1,6 +1,6 @@
 library(fs)
 
-p1 <- test_path("test-aws.R")
+p1 <- test_path("test-count2.R")
 p2 <- test_path("test-last.R")
 
 test_that("wc called for default args", {