+++ NEWS.md
@@ -1,3 +1,5 @@
+# k5 (development version)
+
 # k5 0.0.5
 
 * Add some more basic functions tests.